+++ src/orchestrator/redraw.rs
@@ -5,18 +5,26 @@ use crate::components::BoxComponent;
 
 fn draw_square(
     canvas: &mut Canvas,
-    left: f32,
-    top: f32,
-    right: f32,
-    bottom: f32,
+    left: u32,
+    top: u32,
+    right: u32,
+    bottom: u32,
     color: u32,
 ) {
     canvas.save();
-    let rect = Rect::new(left, top, right, bottom);
+
+    let rect = Rect::new(
+        left as f32,
+        top as f32,
+        right as f32,
+        bottom as f32,
+    );
+
     let mut paint: Paint = Paint::new(
         Color4f::new(0.0, 0.0, 0.0, 0.0),
         None
     );
+
     paint.set_color(color);
     canvas.draw_rect(rect, &paint);
     canvas.restore();
